# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AccountProfileResponseDto": {
        "properties": {
          "accountId": {
            "example": "0f27af0b-48b2-4f1b-b3d4-cd531a0b4458",
            "format": "uuid",
            "type": "string"
          },
          "contactEmail": {
            "example": "Case.Sensitive@example.com",
            "format": "email",
            "maxLength": 254,
            "minLength": 3,
            "pattern": "^[!-~]+$",
            "type": "string"
          },
          "contactPhone": {
            "example": "+14155552671",
            "maxLength": 16,
            "minLength": 3,
            "nullable": true,
            "pattern": "^\\+[1-9][0-9]{1,14}$",
            "type": "string"
          },
          "createdAt": {
            "example": "2026-08-20T16:00:00.000Z",
            "format": "date-time",
            "type": "string"
          },
          "declaredResidencyCountryCode": {
            "enum": [
              "AD",
              "AE",
              "AF",
              "AG",
              "AI",
              "AL",
              "AM",
              "AO",
              "AQ",
              "AR",
              "AS",
              "AT",
              "AU",
              "AW",
              "AX",
              "AZ",
              "BA",
              "BB",
              "BD",
              "BE",
              "BF",
              "BG",
              "BH",
              "BI",
              "BJ",
              "BL",
              "BM",
              "BN",
              "BO",
              "BQ",
              "BR",
              "BS",
              "BT",
              "BV",
              "BW",
              "BY",
              "BZ",
              "CA",
              "CC",
              "CD",
              "CF",
              "CG",
              "CH",
              "CI",
              "CK",
              "CL",
              "CM",
              "CN",
              "CO",
              "CR",
              "CU",
              "CV",
              "CW",
              "CX",
              "CY",
              "CZ",
              "DE",
              "DJ",
              "DK",
              "DM",
              "DO",
              "DZ",
              "EC",
              "EE",
              "EG",
              "EH",
              "ER",
              "ES",
              "ET",
              "FI",
              "FJ",
              "FK",
              "FM",
              "FO",
              "FR",
              "GA",
              "GB",
              "GD",
              "GE",
              "GF",
              "GG",
              "GH",
              "GI",
              "GL",
              "GM",
              "GN",
              "GP",
              "GQ",
              "GR",
              "GS",
              "GT",
              "GU",
              "GW",
              "GY",
              "HK",
              "HM",
              "HN",
              "HR",
              "HT",
              "HU",
              "ID",
              "IE",
              "IL",
              "IM",
              "IN",
              "IO",
              "IQ",
              "IR",
              "IS",
              "IT",
              "JE",
              "JM",
              "JO",
              "JP",
              "KE",
              "KG",
              "KH",
              "KI",
              "KM",
              "KN",
              "KP",
              "KR",
              "KW",
              "KY",
              "KZ",
              "LA",
              "LB",
              "LC",
              "LI",
              "LK",
              "LR",
              "LS",
              "LT",
              "LU",
              "LV",
              "LY",
              "MA",
              "MC",
              "MD",
              "ME",
              "MF",
              "MG",
              "MH",
              "MK",
              "ML",
              "MM",
              "MN",
              "MO",
              "MP",
              "MQ",
              "MR",
              "MS",
              "MT",
              "MU",
              "MV",
              "MW",
              "MX",
              "MY",
              "MZ",
              "NA",
              "NC",
              "NE",
              "NF",
              "NG",
              "NI",
              "NL",
              "NO",
              "NP",
              "NR",
              "NU",
              "NZ",
              "OM",
              "PA",
              "PE",
              "PF",
              "PG",
              "PH",
              "PK",
              "PL",
              "PM",
              "PN",
              "PR",
              "PS",
              "PT",
              "PW",
              "PY",
              "QA",
              "RE",
              "RO",
              "RS",
              "RU",
              "RW",
              "SA",
              "SB",
              "SC",
              "SD",
              "SE",
              "SG",
              "SH",
              "SI",
              "SJ",
              "SK",
              "SL",
              "SM",
              "SN",
              "SO",
              "SR",
              "SS",
              "ST",
              "SV",
              "SX",
              "SY",
              "SZ",
              "TC",
              "TD",
              "TF",
              "TG",
              "TH",
              "TJ",
              "TK",
              "TL",
              "TM",
              "TN",
              "TO",
              "TR",
              "TT",
              "TV",
              "TW",
              "TZ",
              "UA",
              "UG",
              "UM",
              "US",
              "UY",
              "UZ",
              "VA",
              "VC",
              "VE",
              "VG",
              "VI",
              "VN",
              "VU",
              "WF",
              "WS",
              "YE",
              "YT",
              "ZA",
              "ZM",
              "ZW"
            ],
            "example": "CA",
            "maxLength": 2,
            "minLength": 2,
            "type": "string"
          },
          "eligibilityStatus": {
            "enum": [
              "UNKNOWN"
            ],
            "example": "UNKNOWN",
            "type": "string"
          },
          "updatedAt": {
            "example": "2026-08-20T16:00:00.000Z",
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "example": 1,
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "accountId",
          "contactEmail",
          "contactPhone",
          "declaredResidencyCountryCode",
          "eligibilityStatus",
          "version",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "HealthResponseDto": {
        "properties": {
          "service": {
            "example": "crypto-lending-api",
            "type": "object"
          },
          "status": {
            "enum": [
              "ok"
            ],
            "example": "ok",
            "type": "string"
          }
        },
        "required": [
          "service",
          "status"
        ],
        "type": "object"
      },
      "InfrastructureHealthResponseDto": {
        "properties": {
          "checks": {
            "$ref": "#/components/schemas/PublicDependencyChecksDto"
          },
          "status": {
            "enum": [
              "ok",
              "degraded"
            ],
            "example": "ok",
            "type": "string"
          }
        },
        "required": [
          "status",
          "checks"
        ],
        "type": "object"
      },
      "PublicDependencyChecksDto": {
        "properties": {
          "postgres": {
            "$ref": "#/components/schemas/PublicDependencyHealthCheckDto"
          },
          "redis": {
            "$ref": "#/components/schemas/PublicDependencyHealthCheckDto"
          },
          "sqs": {
            "$ref": "#/components/schemas/PublicDependencyHealthCheckDto"
          }
        },
        "required": [
          "postgres",
          "redis",
          "sqs"
        ],
        "type": "object"
      },
      "PublicDependencyHealthCheckDto": {
        "properties": {
          "latencyMs": {
            "example": 2,
            "minimum": 0,
            "type": "number"
          },
          "status": {
            "enum": [
              "up",
              "down"
            ],
            "example": "up",
            "type": "string"
          }
        },
        "required": [
          "status",
          "latencyMs"
        ],
        "type": "object"
      },
      "VersionResponseDto": {
        "properties": {
          "apiVersion": {
            "example": "v1",
            "type": "object"
          },
          "service": {
            "example": "crypto-lending-api",
            "type": "object"
          },
          "serviceVersion": {
            "example": "0.1.0",
            "type": "object"
          }
        },
        "required": [
          "apiVersion",
          "service",
          "serviceVersion"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "description": "Managed identity access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Versioned API for the Crypto Lending platform.",
    "title": "Crypto Lending API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/accounts/me": {
      "get": {
        "operationId": "AccountProfileController_getSelf",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountProfileResponseDto"
                }
              }
            },
            "description": "",
            "headers": {
              "ETag": {
                "description": "Strong account-bound profile version for a later If-Match request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid verified principal"
          },
          "404": {
            "description": "No profile exists for the authenticated account"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read the authenticated account profile",
        "tags": [
          "accounts"
        ]
      },
      "patch": {
        "operationId": "AccountProfileController_updateSelf",
        "parameters": [
          {
            "description": "Strong account-bound ETag returned by the latest successful read",
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "maxProperties": 3,
                "minProperties": 1,
                "properties": {
                  "contactEmail": {
                    "description": "Printable ASCII email; the local part is preserved and the domain is lowercase",
                    "example": "Case.Sensitive@example.com",
                    "format": "email",
                    "maxLength": 254,
                    "minLength": 3,
                    "pattern": "^[!-~]+$",
                    "type": "string"
                  },
                  "contactPhone": {
                    "description": "Canonical E.164 syntax or null to clear the unverified contact number",
                    "example": "+14155552671",
                    "maxLength": 16,
                    "minLength": 3,
                    "nullable": true,
                    "pattern": "^\\+[1-9][0-9]{1,14}$",
                    "type": "string"
                  },
                  "declaredResidencyCountryCode": {
                    "description": "Unverified, officially assigned ISO 3166-1 alpha-2 declaration",
                    "enum": [
                      "AD",
                      "AE",
                      "AF",
                      "AG",
                      "AI",
                      "AL",
                      "AM",
                      "AO",
                      "AQ",
                      "AR",
                      "AS",
                      "AT",
                      "AU",
                      "AW",
                      "AX",
                      "AZ",
                      "BA",
                      "BB",
                      "BD",
                      "BE",
                      "BF",
                      "BG",
                      "BH",
                      "BI",
                      "BJ",
                      "BL",
                      "BM",
                      "BN",
                      "BO",
                      "BQ",
                      "BR",
                      "BS",
                      "BT",
                      "BV",
                      "BW",
                      "BY",
                      "BZ",
                      "CA",
                      "CC",
                      "CD",
                      "CF",
                      "CG",
                      "CH",
                      "CI",
                      "CK",
                      "CL",
                      "CM",
                      "CN",
                      "CO",
                      "CR",
                      "CU",
                      "CV",
                      "CW",
                      "CX",
                      "CY",
                      "CZ",
                      "DE",
                      "DJ",
                      "DK",
                      "DM",
                      "DO",
                      "DZ",
                      "EC",
                      "EE",
                      "EG",
                      "EH",
                      "ER",
                      "ES",
                      "ET",
                      "FI",
                      "FJ",
                      "FK",
                      "FM",
                      "FO",
                      "FR",
                      "GA",
                      "GB",
                      "GD",
                      "GE",
                      "GF",
                      "GG",
                      "GH",
                      "GI",
                      "GL",
                      "GM",
                      "GN",
                      "GP",
                      "GQ",
                      "GR",
                      "GS",
                      "GT",
                      "GU",
                      "GW",
                      "GY",
                      "HK",
                      "HM",
                      "HN",
                      "HR",
                      "HT",
                      "HU",
                      "ID",
                      "IE",
                      "IL",
                      "IM",
                      "IN",
                      "IO",
                      "IQ",
                      "IR",
                      "IS",
                      "IT",
                      "JE",
                      "JM",
                      "JO",
                      "JP",
                      "KE",
                      "KG",
                      "KH",
                      "KI",
                      "KM",
                      "KN",
                      "KP",
                      "KR",
                      "KW",
                      "KY",
                      "KZ",
                      "LA",
                      "LB",
                      "LC",
                      "LI",
                      "LK",
                      "LR",
                      "LS",
                      "LT",
                      "LU",
                      "LV",
                      "LY",
                      "MA",
                      "MC",
                      "MD",
                      "ME",
                      "MF",
                      "MG",
                      "MH",
                      "MK",
                      "ML",
                      "MM",
                      "MN",
                      "MO",
                      "MP",
                      "MQ",
                      "MR",
                      "MS",
                      "MT",
                      "MU",
                      "MV",
                      "MW",
                      "MX",
                      "MY",
                      "MZ",
                      "NA",
                      "NC",
                      "NE",
                      "NF",
                      "NG",
                      "NI",
                      "NL",
                      "NO",
                      "NP",
                      "NR",
                      "NU",
                      "NZ",
                      "OM",
                      "PA",
                      "PE",
                      "PF",
                      "PG",
                      "PH",
                      "PK",
                      "PL",
                      "PM",
                      "PN",
                      "PR",
                      "PS",
                      "PT",
                      "PW",
                      "PY",
                      "QA",
                      "RE",
                      "RO",
                      "RS",
                      "RU",
                      "RW",
                      "SA",
                      "SB",
                      "SC",
                      "SD",
                      "SE",
                      "SG",
                      "SH",
                      "SI",
                      "SJ",
                      "SK",
                      "SL",
                      "SM",
                      "SN",
                      "SO",
                      "SR",
                      "SS",
                      "ST",
                      "SV",
                      "SX",
                      "SY",
                      "SZ",
                      "TC",
                      "TD",
                      "TF",
                      "TG",
                      "TH",
                      "TJ",
                      "TK",
                      "TL",
                      "TM",
                      "TN",
                      "TO",
                      "TR",
                      "TT",
                      "TV",
                      "TW",
                      "TZ",
                      "UA",
                      "UG",
                      "UM",
                      "US",
                      "UY",
                      "UZ",
                      "VA",
                      "VC",
                      "VE",
                      "VG",
                      "VI",
                      "VN",
                      "VU",
                      "WF",
                      "WS",
                      "YE",
                      "YT",
                      "ZA",
                      "ZM",
                      "ZW"
                    ],
                    "example": "CA",
                    "maxLength": 2,
                    "minLength": 2,
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountProfileResponseDto"
                }
              }
            },
            "description": "",
            "headers": {
              "ETag": {
                "description": "Strong account-bound profile version after this update",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Profile fields are invalid or the update is empty"
          },
          "401": {
            "description": "Missing or invalid verified principal"
          },
          "404": {
            "description": "No profile exists for the authenticated account"
          },
          "412": {
            "description": "If-Match is malformed, mismatched, or stale"
          },
          "428": {
            "description": "If-Match is required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update permitted fields on the authenticated account profile",
        "tags": [
          "accounts"
        ]
      }
    },
    "/api/v1/health": {
      "get": {
        "operationId": "HealthController_getHealth",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Report API process health",
        "tags": [
          "system"
        ]
      }
    },
    "/api/v1/health/dependencies": {
      "get": {
        "operationId": "InfrastructureHealthController_checkDependencies",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfrastructureHealthResponseDto"
                }
              }
            },
            "description": "All required dependencies are reachable and correctly configured."
          },
          "429": {
            "description": "This replica has exhausted its bounded readiness request budget."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfrastructureHealthResponseDto"
                }
              }
            },
            "description": "At least one required dependency is unavailable or misconfigured."
          }
        },
        "summary": "Report required dependency readiness",
        "tags": [
          "system"
        ]
      }
    },
    "/api/v1/version": {
      "get": {
        "operationId": "VersionController_getVersion",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Report API contract and service versions",
        "tags": [
          "system"
        ]
      }
    }
  },
  "servers": [],
  "tags": []
}